[Chore][CI] fold trust-routing test into test_reclaim_action; drop remaining policy tests
The structural tests over reclaim wiring re-stated the workflow diff
without adding signal; the trust-routing assertion is kept because a
misroute fails silently (no CI signal). It moves to the existing CI
wiring test file so no dedicated policy test file remains.
