+++ tests/test_reclaim_action.py
@@ -2,7 +2,9 @@
 
 Covers the sentinel-repair + atomic-trim primitives that protect caches
 whose consumers assume "directory exists => contents complete" (the
-tilelang autotuner cache in particular).
+tilelang autotuner cache in particular), plus the gpu-smoke security
+trust-routing contract (a misroute there fails silently — no CI signal —
+so a structural assertion is the only automated guard).
 
 Required cases:
   - half_dead       : atomic subdir with files but no sentinel is removed
@@ -276,3 +278,48 @@ def test_trim_files_removes_old_files_but_leaves_atomic_roots_alone(
 
 def test_trim_files_tolerates_missing_root(tmp_path: Path) -> None:
     _run("trim-files", "7", str(tmp_path / "nope"))
+
+
+# ---------------------------------------------------------------------------
+# gpu-smoke security trust routing
+# ---------------------------------------------------------------------------
+
+GPU_SMOKE_WORKFLOW = REPO_ROOT / ".github" / "workflows" / "gpu-smoke.yml"
+
+
+def test_security_policy_routes_trust_by_collaborator_permission() -> None:
+    """The security-policy step must derive is_fork from the PR author's collaborator
+    permission (write/maintain/admin -> trusted), NOT author_association, and fail closed
+    to the fork pool on any lookup failure. The same is_fork drives runs-on and the
+    trusted-action ref selection."""
+    import yaml
+
+    wf = yaml.safe_load(GPU_SMOKE_WORKFLOW.read_text())
+    policy_job = wf["jobs"]["security-policy"]
+    run_steps = [s for s in policy_job["steps"] if "run" in s and s.get("id") == "policy"]
+    assert run_steps, "expected a 'policy' step in security-policy job"
+    step = run_steps[0]
+    script = step["run"]
+    env = step["env"]
+
+    # Trust is keyed off the collaborator-permission endpoint, not author_association.
+    assert "AUTHOR_ASSOC" not in env, "author_association must no longer drive trust"
+    assert "AUTHOR_ASSOC" not in script
+    assert "collaborators/${PR_AUTHOR}/permission" in script, (
+        "is_fork must be derived from the collaborator-permission endpoint"
+    )
+    # Only write/maintain/admin are trusted; the catch-all fails closed to the fork pool.
+    assert "admin|maintain|write" in script
+    assert 'is_fork="false"' in script  # trusted branch
+    assert 'is_fork="true"' in script  # fail-closed / external branch
+    # PR_AUTHOR must be plumbed from the PR author login.
+    assert "PR_AUTHOR" in env, "PR_AUTHOR must be plumbed via env:"
+    assert "pull_request.user.login" in env["PR_AUTHOR"]
+
+    # runs-on and the trusted-action ref must both consume the same is_fork output.
+    gpu_job = wf["jobs"]["gpu-smoke"]
+    assert "needs.security-policy.outputs.is_fork" in str(gpu_job["runs-on"])
+    ref_step = next(
+        s for s in gpu_job["steps"] if (s.get("name") or "").startswith("Checkout trusted actions")
+    )
+    assert "needs.security-policy.outputs.is_fork" in str(ref_step["with"]["ref"])
